# frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectionApproval": {
        "properties": {
          "approved_by": {
            "default": "technician",
            "maxLength": 120,
            "minLength": 1,
            "title": "Approved By",
            "type": "string"
          }
        },
        "title": "ConnectionApproval",
        "type": "object"
      },
      "CustomerSystem": {
        "properties": {
          "customer_id": {
            "title": "Customer Id",
            "type": "integer"
          },
          "system": {
            "$ref": "#/components/schemas/SystemInfo"
          },
          "ticket_id": {
            "title": "Ticket Id",
            "type": "integer"
          }
        },
        "required": [
          "ticket_id",
          "customer_id",
          "system"
        ],
        "title": "CustomerSystem",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntegrationRequest": {
        "properties": {
          "activity_draft_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Draft Id"
          },
          "activity_payload": {
            "additionalProperties": true,
            "title": "Activity Payload",
            "type": "object"
          },
          "attempts": {
            "default": 0,
            "title": "Attempts",
            "type": "integer"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "phoenix_activity_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phoenix Activity Id"
          },
          "request_type": {
            "title": "Request Type",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/IntegrationRequestStatus"
          },
          "ticket_id": {
            "title": "Ticket Id",
            "type": "integer"
          },
          "ticket_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TicketStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "ticket_id",
          "request_type",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "IntegrationRequest",
        "type": "object"
      },
      "IntegrationRequestStatus": {
        "enum": [
          "pending",
          "processing",
          "activity_created",
          "completed",
          "failed",
          "dead_letter"
        ],
        "title": "IntegrationRequestStatus",
        "type": "string"
      },
...
      "Run": {
        "properties": {
          "activity_draft": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Draft"
          },
          "current_hypotheses": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Current Hypotheses",
            "type": "array"
          },
          "customer_system_snapshot": {
            "additionalProperties": true,
            "title": "Customer System Snapshot",
            "type": "object"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "pending_step": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Step"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "ticket_id": {
            "title": "Ticket Id",
            "type": "integer"
          },
          "ticket_snapshot": {
            "additionalProperties": true,
            "title": "Ticket Snapshot",
            "type": "object"
          },
          "validation_result": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Result"
          }
        },
        "required": [
          "id",
          "ticket_id",
          "status",
          "started_at",
          "ticket_snapshot",
          "customer_system_snapshot"
        ],
        "title": "Run",
        "type": "object"
      },
      "RunAbort": {
        "properties": {
          "aborted_by": {
            "default": "technician",
            "maxLength": 120,
            "minLength": 1,
            "title": "Aborted By",
            "type": "string"
          },
          "reason": {
            "default": "Aborted by technician.",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "RunAbort",
        "type": "object"
      },
      "RunCreate": {
        "properties": {
          "ticket_id": {
            "exclusiveMinimum": 0.0,
            "title": "Ticket Id",
            "type": "integer"
          }
        },
        "required": [
          "ticket_id"
        ],
        "title": "RunCreate",
        "type": "object"
      },
...
      "RunRetry": {
        "properties": {
          "reason": {
            "default": "Retry requested by technician.",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "requested_by": {
            "default": "technician",
            "maxLength": 120,
            "minLength": 1,
            "title": "Requested By",
            "type": "string"
          }
        },
        "title": "RunRetry",
        "type": "object"
      },
      "RunStatus": {
        "enum": [
          "awaiting_connection_approval",
          "investigating",
          "planning",
          "awaiting_step_approval",
          "executing",
          "fixing",
          "validating",
          "ready_for_activity",
          "submitted",
          "aborted",
          "failed"
        ],
        "title": "RunStatus",
        "type": "string"
      },
      "StatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusUpdate",
        "type": "object"
      },
...
      "SystemInfo": {
        "properties": {
          "ip": {
            "title": "Ip",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "os": {
            "title": "Os",
            "type": "string"
          },
          "port": {
            "title": "Port",
            "type": "integer"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "ip",
          "port",
          "username",
          "os"
        ],
        "title": "SystemInfo",
        "type": "object"
      },
      "Ticket": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "customer_id": {
            "title": "Customer Id",
            "type": "integer"
          },
          "customer_name": {
            "title": "Customer Name",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "priority": {
            "title": "Priority",
            "type": "string"
          },
          "sla_due_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Due At"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "priority",
          "status",
          "customer_id",
          "customer_name"
        ],
        "title": "Ticket",
        "type": "object"
      },
      "TicketStatus": {
        "enum": [
          "OPEN",
          "PENDING",
          "DONE"
        ],
        "title": "TicketStatus",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backend control plane for Phoenix ERP, technician approvals, SSH execution, and audit state.",
    "title": "techbold AI Service Desk Autopilot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs": {
      "post": {
        "operationId": "start_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Run"
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run"
      }
    },
    "/api/runs/{run_id}/abort": {
      "post": {
        "operationId": "abort_run_api_runs__run_id__abort_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunAbort"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Abort Run"
      }
    },
...
    "/api/runs/{run_id}/connect/approve": {
      "post": {
        "operationId": "approve_run_connection_api_runs__run_id__connect_approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionApproval"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Run Connection"
      }
    },
...
    "/api/runs/{run_id}/integration-requests": {
      "get": {
        "operationId": "get_run_integration_requests_api_runs__run_id__integration_requests_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IntegrationRequest"
                  },
                  "title": "Response Get Run Integration Requests Api Runs  Run Id  Integration Requests Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Integration Requests"
      }
    },
    "/api/runs/{run_id}/integration-requests/{integration_request_id}": {
      "get": {
        "operationId": "get_run_integration_request_api_runs__run_id__integration_requests__integration_request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "integration_request_id",
            "required": true,
            "schema": {
              "title": "Integration Request Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationRequest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Integration Request"
      }
    },
...
    "/api/runs/{run_id}/retry": {
      "post": {
        "operationId": "retry_run_api_runs__run_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRetry"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Run"
      }
    },
...
    "/api/tickets": {
      "get": {
        "operationId": "list_tickets_api_tickets_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TicketStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "date",
              "pattern": "^(date|priority|status)$",
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Ticket"
                  },
                  "title": "Response List Tickets Api Tickets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tickets"
      }
    },
    "/api/tickets/{ticket_id}": {
      "get": {
        "operationId": "get_ticket_api_tickets__ticket_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "title": "Ticket Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ticket"
      }
    },
    "/api/tickets/{ticket_id}/customer-system": {
      "get": {
        "operationId": "get_customer_system_api_tickets__ticket_id__customer_system_get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "title": "Ticket Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSystem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Customer System"
      }
    },
    "/api/tickets/{ticket_id}/status": {
      "patch": {
        "operationId": "set_ticket_status_api_tickets__ticket_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "title": "Ticket Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ticket"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Ticket Status"
      }
    },
...
  }
}